pyproject.toml
[project]
name = "tdm-geo-analysiszones-releases"
version = "0.1.0"
description = "Official release repository for Wasatch Front Travel Demand Model Analysis Zones (TAZ/MAZ)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arcgis>=2.4.3",
    "fiona>=1.10.1",
    "geopandas>=1.1.3",
    "pandas>=3.0.2",
    "pyogrio>=0.12.1",
    "pyshp<3",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/analysiszones"]
